scrapers/calendar_scraper/requirements.txt
# Core
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Database
python-dotenv>=1.0.0
oracledb>=1.4.0

# Scheduling (optional — for local cron-like runs)
schedule>=1.2.0
